[ad-mxfp4-moe] Drop dead load-hook + move sharding hook + clarify naming
Deletes unused make_mxfp4_trtllm_load_hook + _get_default_dist_info +
_hook_dist_info_fn and stale V4-plan / make_mxfp4_ep_slice_load_hook
docstring references. Relocates make_mxfp4_sharding_load_hook to
transform/library/mxfp4_moe.py next to its only caller. Renames
swizzle_moe_mxfp4_weights{.py,()} -> prepare_trtllm_gen_moe_mxfp4_weights{.py,()}
and PreparedMXFP4Weights -> TRTLLMGenMXFP4MoEWeights.
